pyproject.toml
[project]
name = "animalitos"
version = "0.1.0"
description = ""
authors = [
    {name = "simon",email = "simondjv2016@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "sqlmodel (>=0.0.22,<0.0.23)",
    "pydantic-settings (>=2.8.0,<3.0.0)",
    "strawberry-graphql[fastapi] (>=0.260.2,<0.261.0)",
    "boto3 (>=1.36.26,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false 
